+++ cartridges/int_cybs_sfra_base/cartridge/scripts/hooks/payment/processor/payments_credit_form_processor.js
@@ -31,7 +31,7 @@ function processForm(req, paymentForm, viewFormData) {
             case 'discover':
                 correctCardType = 'Discover';
                 break;
-            case 'dinersclub':
+            case 'diners-club':
                 correctCardType = 'DinersClub';
                 break;
             case 'maestro':
@@ -40,6 +40,18 @@ function processForm(req, paymentForm, viewFormData) {
             case 'jcb':
                 correctCardType = 'JCB';
                 break;
+            case "cartesbancaires":
+                correctCardType = "CartesBancaires";
+                break;
+            case "elo":
+                correctCardType = "Elo";
+                break;
+            case "cup":
+                correctCardType = "China UnionPay";
+                break;
+            case "jcrew":
+                correctCardType = "JCrew";
+                break;     
         }
         // eslint-disable-next-line no-param-reassign
         paymentForm.creditCardFields.cardType.value = correctCardType;